+++ client/src/components/schedule/index.ts
@@ -1,0 +1,3 @@
+export { default as TripInfo } from './TripInfo';
+export { default as TripSchedule } from './TripSchedule';
+export { default as ScheduleAccordion } from './ScheduleAccordion';
